Add jointid_t enum for joints
