+++ src/world_interface/data.cpp
@@ -28,3 +28,30 @@ constexpr float PotentiometerParams::scale() const {
 	return (static_cast<int16_t>(adc_hi) - static_cast<int16_t>(adc_lo))/(mdeg_hi - mdeg_lo);
 }
 } // namespace robot::types
+
+namespace std {
+std::string to_string(robot::types::jointid_t joint) {
+	using robot::types::jointid_t;
+	switch (joint) {
+		case jointid_t::armBase:
+			return "armBase";
+		case jointid_t::shoulder:
+			return "shoulder";
+		case jointid_t::elbow:
+			return "elbow";
+		case jointid_t::forearm:
+			return "forearm";
+		case jointid_t::hand:
+			return "hand";
+		case jointid_t::differentialPitch:
+			return "differentialPitch";
+		case jointid_t::differentialRoll:
+			return "differentialRoll";
+		case jointid_t::drill_arm:
+			return "drillArm";
+		default:
+			// should never happen
+			return "<unknown>";
+	}
+}
+} // namespace std
